testing: add different allocators to different parts of tests
This gives a much clearer picture of what magnitude of memory usage each
part has and will provide (in some sense) an upper bound on what it can
be.

Granted, these are arenas so they don't show total memory usage at any
one point, but will definitely signal clearly when tons of allocations
and frees are happening in a code path.